+++ test_models/print/print.stan
@@ -1,0 +1,8 @@
+parameters {
+   real theta;
+}
+model {
+  theta ~ normal(0, 1);
+  print("Hi from Stan!");
+  print("theta = ", theta);
+}
